clean up duplicate tests
